kernel: enable I2C bus sharing logic for BYT/CHT
On Baytrail and Cherry-Trail (UP Board) SoCs, the P-Unit platform firmware
needs to share access to the PMIC I2C bus.  To avoid a race-condition here,
the Designware I2C driver should request a semaphore from the P-Unit before
attempting to access the bus.  To enable this, the following kernel config
options are required:
- CONFIG_I2C_DESIGNWARE_BAYTRAIL
- CONFIG_IOSF_MBI