kernel: enable reboot on kernel panic for recovery
UP boards are often used in remote headless deployments and a kernel panic
can be a major issue if the board doesn't recover autonomously and becomes
unreachable.  Enabling a watchdog timer can sometimes provide a solution,
but  typically will miss a panic which occurs early during system start-up.

To mitigate that, this commit enables the kernel to issue a hard reboot
5 seconds after a kernel panic occurs in an attempt to bring the board
back to a working state.