kernel: panic and reboot on any kernel warning
Trigger a panic and subsequent reboot if a kernel driver issues a warning.
This is intended to help an UP board to recover from a deadlock which is
known to occur at random during system start-up while requesting an I2C
bus access semaphore from unresponsive P-Unit platform firmware.

Although this issue initially manifests as a kernel warning, it ultimately
leads to the board becoming 'hung' some seconds later when the i915 driver
is loaded.  Currently, the only viable option is to reboot the board when
this occurs.