kernel:up: Preparation to enable new UP Models on Warrior branch
Until now only UP-Xtreme was supported on warrior branch.

This is an initial commit that adds new recipe-kernel patches and to allow
images to be built for up-board, up-squared, up-core and up-core-plus
on warrior.

No additional boards are enabled in this commit, these will follow later.

Tested UP-board boots (for linux-intel) and UP-Xtreme image build.

Signed-off-by: Michael Campion <michael.campion@emutex.com>